NPC Stuff and General Systems
Player Camera Sensitivity felt way too slow. It's a public variable so I'm assuming that it was meant to be changed anyways. I set it to 500 because it felt good to me, but still a little slow.
Set project to 1920x1080p
Imported Text Mesh Pro Essentials

Player Script:
Detects when NPC is in range of the player

GameManager Script:
Accessible by all scripts
Controls universal things like UI.
Tracks information like whether or not we've got the macguffin

UIManager Script:
It's attached to the canvas
Will control UI elements usually using information from the GameManager

TextEffect Script:
Creates a typing sort of text effect so that text doesn't just appear on the screen

I was messing with the PlayerLook Script to make the camera turn towards the NPC that they're interacting with.
When you walk up to the Temp Frog, a F button will appear on your screen. If you press f, "Howdy!" will pop up in the center of your screen.